fix: dockerising and testing, deleting not needed endpoints
