+++ api/main.py
@@ -1,12 +1,12 @@
 from fastapi import FastAPI
 
 from api.configs.Environment import get_environment_variables
-from api.metadata.Tags import Tags
+from api.contact.metadata.Tags import Tags
 from api.contact.models.BaseModel import init
 from api.contact.routers.v1.InvestigateContactsRouter import InvestigateContactRouter
 from api.contact.routers.v1.ContactsRouter import ContactsRouter
-from api.subscriber_api.routers.InvestigateSubscriberContractRouter import InvestigateContractRouter
-from api.subscriber_api.routers.SubscriberContractRouter import SubscriberContractAPIRouter
+from api.subscription.routers.InvestigateSubscriberContractRouter import InvestigateContractRouter
+from api.subscription.routers.SubscriberContractRouter import SubscriberContractAPIRouter
 
 # Application Environment Configuration
 env = get_environment_variables()
@@ -20,7 +20,7 @@ app = FastAPI(
 
 # Add Routers
 app.include_router(ContactsRouter)
-app.include_router(InvestigateContactRouter)
+# app.include_router(InvestigateContactRouter)
 app.include_router(SubscriberContractAPIRouter)
 app.include_router(InvestigateContractRouter)
