--- api/subscriber_api/configs/Database.py
@@ -1,30 +1,0 @@
-from sqlalchemy import create_engine
-from sqlalchemy.orm import scoped_session, sessionmaker
-
-from api.subscriber_api.configs.Environment import get_environment_variables
-
-# Runtime Environment Configuration
-env = get_environment_variables()
-
-# Generate Database URL
-DATABASE_URL = f"{env.DATABASE_DIALECT}://{env.DATABASE_USERNAME}:{env.DATABASE_PASSWORD}@{env.DATABASE_HOSTNAME}:{env.DATABASE_PORT}/{env.DATABASE_NAME}"
-
-
-#DATABASE_URL = "mysql://authuser:authuser@127.0.0.1:3306/testdb"
-
-# Create Database Engine
-Engine = create_engine(
-    DATABASE_URL, echo=False, future=True
-)
-
-SessionLocal = sessionmaker(
-    autocommit=False, autoflush=False, bind=Engine
-)
-
-
-def get_db_connection():
-    db = scoped_session(SessionLocal)
-    try:
-        yield db
-    finally:
-        db.close()
