--- api/subscriber_api/configs/Environment.py
@@ -1,37 +1,0 @@
-from functools import lru_cache
-import os
-
-from pydantic import BaseSettings
-
-
-@lru_cache
-def get_env_filename():
-    runtime_env = os.getenv("ENV")
-    return f".env.{runtime_env}" if runtime_env else ".env"
-
-
-class EnvironmentSettings(BaseSettings):
-    API_VERSION: str
-    APP_NAME: str
-    DATABASE_DIALECT: str
-    DATABASE_HOSTNAME: str
-    DATABASE_NAME: str
-    DATABASE_PASSWORD: str
-    DATABASE_PORT: int
-    DATABASE_USERNAME: str
-    DEBUG_MODE: bool
-    API_AUTH_HOST: str
-    API_AUTH_PORT: str
-    API_AUTH_PATH_INTROSPECT: str
-    API_VERSION_STR: str
-    API_SCHEMA: str
-
-    class Config:
-        env_file = get_env_filename()
-        env_file_encoding = "utf-8"
-
-
-@lru_cache
-def get_environment_variables():
-    return EnvironmentSettings()
-
