--- api/subscriber_api/constant.py
@@ -1,78 +1,0 @@
-class OpenAPIFieldDescription:
-    SUBSCRIBER_TYPE = "This field define the subscriber type"
-    POWER_OF_ENERGY = "This field define the power of energy"
-    STATUS = "This field define the contract's status"
-    PREVIOUS_STATUS = "This field define the contract's previous status"
-    INVOICING_FREQUENCY = "This field define the invoicing frequency of the contract"
-    DELIVERY_POINT = "This field define contract's delivery point.A contract can be associate with one delivery point"
-    METRIC_NUMBER = "This field define the number of the electric meter"
-    CUSTOMER_ID = "This field define a customer id"
-    CONTRACT_OPENING_DATE = "This field define a date when the contract has been opened"
-    CONTRACT_CLOSING_DATE = "This field define a date when the contract has been closed (resigned)"
-    CONTRACT_CREATED_AT = "This field define a date when the contract has been created"
-    CONTRACT_UPDATED_AT = "This field define a date when the contract's information has been updated"
-    IS_ACTIVATED = "This field define a date when the contract has been activated - True the contract is activated" \
-                   " False contract is not activated yet "
-    CONTRACT_UID = "This field define the unique number of contract"
-    AGENCY_NAME = "This field define the name of agency"
-    AGENCY_CODE = "This field define the code of agency"
-    AGENCY_TEL = "This field define the phone number of agency"
-    AGENCY_EMAIL = "This field define the email of agency"
-    CONNECTION_POINT = "This is a dict containing the information of connection point"
-    ADDRESS = "This field contain the address information"
-    DELIVERY_POINT_CREATED_AT = "This field define a date when the delivery point has been created"
-    DELIVERY_POINT_UPDATED_AT = "This field define a date when the information of  delivery point has been updated"
-    CITY_NAME = "Name of city"
-    ADMIN_REGION_NAME = "This field define the name of administrative region"
-    NATURAL_REGION_NAME = "This field define the name of natural region"
-    PREFECTURE_NAME = "This field define the name of prefecture"
-    TRANSFORMER_INFO = "This field is a dict and contain the information of transformer"
-    ENERGY_DEPARTURE = "This field is a dict and contain the information of energy departure"
-    FREQUENCY_NAME = "This field contain the name of frequency of payment"
-    FREQUENCY_SHORTNAME = "This field contain the shortname of frequency of payment"
-    CONTRACT_INFO_INPUT_SUBSCRIBER_TYPE = "The id of subscriber type from the ref_subscriber table"
-    CONTRACT_INFO_INPUT_STATUS = "The id of status from ref_status table"
-    CONTRACT_INFO_INPUT_PREVIOUS_STATUS = "The id of the previous status of the contract, this is mapped form ref " \
-                                          "status "
-    CONTRACT_INFO_INPUT_LEVEL = "The level of subscriber contract from the ref_level table"
-    CONTRACT_INFO_INPUT_DELIVERY_POINT = "The id of subscriber contract's delivery point or the delivery point number"
-    CONTRACT_INFO_INPUT_PREVIOUS_LEVEL = "The previous level of subscriber contract from the ref_level table"
-    CONTRACT_INFO_INPUT_AGENCY = "The id of agency"
-    PAYMENT_MODE_NAME = "The name of payment mode"
-    ENERGY_DEPARTURE_INFO = "This is a dict"
-    CREATED_AT = "The creation date"
-    CONTRACT_INFO_OUTPUT_LEVEL = "This is a dict containing a level information"
-    CONTRACT_INFO_OUTPUT_PREVIOUS_LEVEL = "This is a dict containing a previous level information"
-    AGENCY = "Contain a full information about agency"
-    CONTRACT_SCHEMA_INFO = "Content a basic information about the incoming contract while saving or updating"
-    CONTRACT_OUTPUT_INFO = "Content a basic information about the contract"
-    CONTRACT_INFO_LEVEL_NAME = "Contract level name"
-    SUBSCRIPTION_CODE = "Subscription's code"
-    SUBSCRIPTION_NAME = "Name of subscription"
-    PAYMENT_DEADLINE = "Payment deadline"
-    PAYMENT_UNIT_DEADLINE = "Payment deadline unit may be in day in month etc ..."
-    SUBSCRIBED_POWER = "Power of the energy that the contact subscribed "
-    DELIVERY_NUMBER = "Delivery point number"
-    HomeInfos = "This field content whole information of home"
-    SURFACE = "Home surface"
-    OCCUPIER_NUMBER = "Number of person living at home"
-    EQUIPMENT = "Equipment using at home"
-    AMOUNT_EQUIPMENT = "Amount of equipment"
-    EQUIPMENT_LIST = "Equipment list"
-    IS_BLOCKED_PAYED = "This show whether the pricing should be saved in subscriber microservice or not," \
-                       "if the pricing will be saved, so the billing microservice will use it, and calculate " \
-                       "amount of money to pay "
-    PRICING = "Indicate the pricing of the contract"
-    SLICE_NAME = "This field indicate the name of slice"
-    LOWER_INDEX = "This indicate the lower index of dunning"
-    UPPER_INDEX = "This field indicate the upper index of dunning"
-    UNIT_PRICE = "This field indicate the unit price of power energy"
-    DUNNING = "This field contain the information about the number of time the pricing was relaunched"
-    DUNNING_NAME = "name of dunning"
-    DUNNING_RANK = "The number of time the contract was relaunched"
-    DELAY_PENALTY_RATE = "This field contain the rate of penalty to apply while calculating the billing"
-    TRACKING_TYPE = "Type of tracking example tracking_type {\"code\": 2,  \"name\":\"post-payé\"}"
-    TRACKING_CODE = "The code of tracking"
-    TRACKING_NAME = "Name of tracking"
-
-
