--- api/subscriber_api/models/BaseModel.py
@@ -1,10 +1,0 @@
-from sqlalchemy.ext.declarative import declarative_base
-
-from api.subscriber_api.configs.Database import Engine
-
-# Base Entity Model Schema
-EntityMeta = declarative_base()
-
-
-def init():
-    EntityMeta.metadata.create_all(bind=Engine)
