--- api/subscriber_api/utilis/Status.py
@@ -1,3 +1,0 @@
-
-class Status:
-    INITIAL = "Initial"
