+++ api/subscription/services/GuidGenerator.py
