+++ api/subscription/utilis/Status.py
@@ -1,0 +1,10 @@
+
+class ContractStatus:
+    INITIAL = "Initial"
+    EDITED = "Avenant"
+    PENDED = "Suspendu"
+    CLOSED = "Clôturé"
+    CREATED = "Créé"
+    INCOMPLETE = "Incomplet"
+    WAITING_FOR_ACTIVATION = "En attente d’activation"
+
